B1.2.4

+++ radiopadre_client/backends/singularity.py
@@ -1,7 +1,7 @@
 import os, subprocess, sys, time, re, calendar
-import datetime
+import datetime, getpass
 
-from iglesia.utils import message, warning, error, bye, make_dir, make_radiopadre_dir, shell, DEVNULL, ff, INPUT, check_output
+from iglesia.utils import message, warning, error, bye, make_dir, make_radiopadre_dir, shell, DEVNULL, INPUT, check_output
 from radiopadre_client import config
 
 singularity = None
@@ -38,7 +38,7 @@ def kill_sessions(session_dict, session_ids):
 
 def get_singularity_image(docker_image):
     dir = config.SINGULARITY_IMAGE_DIR or os.environ.get('RADIOPADRE_SINGULARITY_IMAGE_DIR') or iglesia.RADIOPADRE_DIR
-    return "{}/{}.singularity.img".format(dir, docker_image.replace("/", "_"))
+    return "{}/{}.simg".format(dir, docker_image.replace("/", "_"))
 
 def update_installation(rebuild=False, docker_pull=True):
     global docker_image
@@ -50,16 +50,16 @@ def update_installation(rebuild=False, docker_pull=True):
 
     # clearly true if no image
     if not os.path.exists(singularity_image):
-        if config.SINGULARITY_AUTO_BUILD:
-            message(ff("Singularity image {singularity_image} does not exist"))
+        if config.SINGULARITY_AUTO_BUILD or config.SINGULARITY_REBUILD:
+            message(f"Singularity image {singularity_image} does not exist, will build")
             build_image = True
         else:
-            error(ff("Singularity image {singularity_image} does not exist, and auto-build is disabled"))
-            bye(ff("Re-run with --singularity-auto-build to proceed"))
+            error(f"Singularity image {singularity_image} does not exist, and auto-build is disabled")
+            bye(f"Re-run with --singularity-rebuild or --singularity-auto-build to proceed")
     # also true if rebuild forced by flags or config or command line
     elif rebuild or config.SINGULARITY_REBUILD:
         config.SINGULARITY_AUTO_BUILD = build_image = True
-        message(ff("--singularity-rebuild specified, removing singularity image {singularity_image}"))
+        message(f"--singularity-rebuild specified, removing singularity image {singularity_image}")
 
     # pull down docker image first
     if has_docker and docker_pull:
@@ -70,8 +70,8 @@ def update_installation(rebuild=False, docker_pull=True):
         if has_docker:
             # check timestamp of docker image
             docker_image_time = None
-            output = check_output(ff("{has_docker} image inspect {docker_image} -f '{{{{ .Created }}}}'")).strip()
-            message(ff("  docker image timestamp is {output}"))
+            output = check_output(f"{has_docker} image inspect {docker_image} -f '{{{{ .Created }}}}'").strip()
+            message(f"  docker image timestamp is {output}")
             # in Python 3.7 we have datetime.fromisoformat(date_string), but for now we muddle:
             match = output and re.match("(^.*)[.](\d+)Z", output)
             if match:
@@ -83,23 +83,26 @@ def update_installation(rebuild=False, docker_pull=True):
             sing_image_time = datetime.datetime.utcfromtimestamp(os.path.getmtime(singularity_image))
             message("  singularity image timestamp is {}".format(sing_image_time.isoformat()))
             if docker_image_time is None:
-                warning(ff("can't parse docker image timestamp '{output}', rebuilding {singularity_image} just in case"))
+                warning(f"can't parse docker image timestamp '{output}', rebuilding {singularity_image} just in case")
                 build_image = True
             elif docker_image_time > sing_image_time:
-                warning(ff("rebuilding outdated singularity image {singularity_image}"))
+                warning(f"rebuilding outdated singularity image {singularity_image}")
                 build_image = True
             else:
-                message(ff("singularity image {singularity_image} is up-to-date"))
+                message(f"singularity image {singularity_image} is up-to-date")
         else:
-            message(ff("--update specified but no docker access, assuming {singularity_image} is up-to-date"))
+            message(f"--update specified, will force-rebuild {singularity_image} from registry")
+            message(f"(If docker was available, we could check timestamps to see if this step is necessary. But we can't.)")
+            build_image = True
+
     # now build if needed
     if build_image:
-        warning(ff("Rebuilding singularity image from docker://{docker_image}"))
-        warning(ff("  (This may take a few minutes....)"))
-        singularity_image_new = singularity_image + ".new.img"
+        warning(f"Rebuilding singularity image from docker://{docker_image}")
+        warning(f"  (This may take a few minutes....)")
+        singularity_image_new = os.path.splitext(singularity_image)[0] + ".new.simg"
         if os.path.exists(singularity_image_new):
             os.unlink(singularity_image_new)
-        cmd = [singularity, "build", singularity_image_new, ff("docker://{docker_image}")]
+        cmd = [singularity, "build", singularity_image_new, f"docker://{docker_image}"]
         message("running " + " ".join(cmd))
         try:
             subprocess.check_call(cmd)
@@ -112,11 +115,18 @@ def update_installation(rebuild=False, docker_pull=True):
                     raise
             else:
                 raise
+        if not os.path.exists(singularity_image_new):
+            if config.IGNORE_UPDATE_ERRORS and os.path.exists(singularity_image):
+                warning("singularity build did not return a error, but the new image did not appear")
+                warning("since --ignore-update-errors is set, we're proceeding with the old image")
+            else:
+                bye("singularity build did not return a error, but the new image did not appear")
         # move old image
-        message(ff("Build successful, renaming to {singularity_image}"))
-        os.rename(singularity_image_new, singularity_image)
+        else:
+            message(f"Build successful, renaming to {singularity_image}")
+            os.rename(singularity_image_new, singularity_image)
     else:
-        message(ff("Using existing radiopadre singularity image {singularity_image}"))
+        message(f"Using existing radiopadre singularity image {singularity_image}")
 
     # not supported with Singularity
     config.CONTAINER_PERSIST = False
@@ -124,36 +134,45 @@ def update_installation(rebuild=False, docker_pull=True):
     # config.CONTAINER_DEBUG = False
 
 
-def start_session(container_name, selected_ports, userside_ports, notebook_path, browser_urls):
-    from iglesia import ABSROOTDIR, LOCAL_SESSION_DIR, SHADOW_SESSION_DIR
+def start_session(container_name, selected_ports, userside_ports, notebook_path, browser_urls, run_browser=False):
+    from iglesia import ABSROOTDIR, SHADOW_SESSION_DIR
     radiopadre_dir = make_radiopadre_dir()
     docker_local = make_dir(radiopadre_dir + "/.docker-local")
     js9_tmp = make_dir(radiopadre_dir + "/.js9-tmp")
     session_info_dir = get_session_info_dir(container_name)
-
-    # message(ff("Container name: {container_name}"))  # remote script will parse it
-
-    os.environ["RADIOPADRE_CONTAINER_NAME"] = container_name
-    os.environ["XDG_RUNTIME_DIR"] = ""
-    docker_opts = ["--workdir", ABSROOTDIR]
-    # setup mounts for work dir and home dir, if needed
     homedir = os.path.expanduser("~")
+
+    docker_opts = []
+
+    if homedir != ABSROOTDIR:
+        docker_opts += ["--workdir", ABSROOTDIR]
+
+    # add other options
+    if config.SINGULARITY_OPTIONS:
+        opts = config.SINGULARITY_OPTIONS.split(",")
+        if not opts[0].startswith("-"):
+            opts[0] = '--' + opts[0]
+        docker_opts += opts
+
+    # setup mounts for work dir and home dir, if needed
     docker_opts += [
         "-B", "{}:{}{}".format(ABSROOTDIR, ABSROOTDIR, ""), # ":ro" if orig_rootdir else ""),
         "-B", "{}:{}".format(radiopadre_dir, radiopadre_dir),
         # hides /home/user/.local, which if exposed, can confuse jupyter and ipython
         "-B", "{}:{}".format(docker_local, os.path.realpath(os.path.join(homedir, ".local"))),
         # mount session info directory (needed to serve e.g. js9prefs.js)
-        "-B", "{}:{}".format(session_info_dir, LOCAL_SESSION_DIR),
         "-B", "{}:{}".format(session_info_dir, SHADOW_SESSION_DIR),
         # mount a writeable tmp dir for the js9 install -- needed by js9helper
         "-B", "{}:/.radiopadre/venv/js9-www/tmp".format(js9_tmp),
+        # mount home session dir
+        "-B", f"{homedir}/.radiopadre-session:{homedir}/.radiopadre-session"
     ]
     if config.CONTAINER_DEV:
         if os.path.isdir(config.CLIENT_INSTALL_PATH):
             docker_opts += ["-B", "{}:/radiopadre-client".format(config.CLIENT_INSTALL_PATH)]
         if os.path.isdir(config.SERVER_INSTALL_PATH):
             docker_opts += ["-B", "{}:/radiopadre".format(config.SERVER_INSTALL_PATH)]
+
     # if not config.CONTAINER_DEBUG:
     #     command = [singularity, "instance.start"] + docker_opts + \
     #               [singularity_image, container_name]
@@ -168,11 +187,20 @@ def start_session(container_name, selected_ports, userside_ports, notebook_path,
     # build up command-line arguments
     docker_opts += _collect_runscript_arguments(container_ports + userside_ports)
 
+    # add environment
+    docker_opts += ["--env", f"JUPYTER_DATA_DIR=/tmp/{getpass.getuser()}-jupyter",
+                    "--env", f"IPYTHONDIR=/tmp/{getpass.getuser()}-ipython",
+                    "--env", f"XDG_RUNTIME_DIR="]
+    os.environ["RADIOPADRE_CONTAINER_NAME"] = container_name
+    for name, value in os.environ.items():
+        if name.startswith("RADIOPADRE_"):
+            docker_opts += ["--env", f"{name}={value}"]
+
     if notebook_path:
         docker_opts.append(notebook_path)
 
-    _run_container(container_name, docker_opts, jupyter_port=selected_ports[0], browser_urls=browser_urls,
-                   singularity=True)
+    _run_container(container_name, docker_opts, jupyter_port=selected_ports[0], 
+                    browser_urls=browser_urls, run_browser=run_browser, singularity=True)
 
     if config.NBCONVERT:
         return
@@ -200,4 +228,4 @@ def start_session(container_name, selected_ports, userside_ports, notebook_path,
 
 def kill_container(name):
     singularity_image = get_singularity_image(config.DOCKER_IMAGE)
-    shell(ff("{singularity} instance.stop {singularity_image} {name}"))
+    shell(f"{singularity} instance.stop {singularity_image} {name}")
